Introduce Pre request interceptor
* Token refresh improvement

* 

* Add logs

* Replace dev
