+++ go.mod
@@ -19,7 +19,7 @@ require (
 	gopkg.in/yaml.v2 v2.2.2
 )
 
-//replace github.com/jfrog/jfrog-client-go => github.com/jfrog/jfrog-client-go dev
+replace github.com/jfrog/jfrog-client-go => github.com/jfrog/jfrog-client-go dev
 
 //replace github.com/jfrog/gocmd => github.com/jfrog/gocmd master
